[FIX] 팀 아이콘 이미지 없을 시 발생하는 버그 수정

+++ src/main/java/com/tiki/server/external/util/AwsHandler.java
@@ -6,6 +6,7 @@ import static com.tiki.server.external.message.ErrorCode.*;
 import static com.tiki.server.external.constant.ExternalConstant.FILE_DELIMITER;
 
 import java.time.Duration;
+import java.util.Objects;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Value;
@@ -47,6 +48,9 @@ public class AwsHandler {
 	}
 
 	public String getDownloadPreSignedUrl(final String fileKey) {
+		if (Objects.isNull(fileKey)) {
+			return "";
+		}
 		try {
 			S3Presigner preSigner = awsConfig.getS3PreSigner();
 			GetObjectRequest getObjectRequest = createGetObjectRequest(fileKey);
@@ -58,6 +62,7 @@ public class AwsHandler {
 	}
 
 	public void deleteFile(final String request) {
+		if (request.equals("file/갓슈.png")) return;	// 임시 로직
 		try {
 			S3Client s3Client = awsConfig.getS3Client();
 			s3Client.deleteObject((DeleteObjectRequest.Builder builder) ->
